+++ bot_jobs/census_data_processing/th-census-data.csv
@@ -1,0 +1,77 @@
+กรุงเทพมหานคร,"5,588,222"
+สมุทรปราการ,"1,351,479"
+นนทบุรี,"1,276,745"
+ปทุมธานี,"1,176,412"
+พระนครศรีอยุธยา,"819,088"
+อ่างทอง,"276,584"
+ลพบุรี,"742,928"
+สิงห์บุรี,"205,898"
+ชัยนาท,"322,477"
+สระบุรี,"643,828"
+ชลบุรี,"1,566,885"
+ระยอง,"741,524"
+จันทบุรี,"535,559"
+ตราด,"228,536"
+ฉะเชิงเทรา,"720,718"
+ปราจีนบุรี,"493,670"
+นครนายก,"260,081"
+สระแก้ว,"560,925"
+ราชบุรี,"869,313"
+กาญจนบุรี,"891,976"
+สุพรรณบุรี,"838,628"
+นครปฐม,"920,729"
+สมุทรสาคร,"586,199"
+สมุทรสงคราม,"192,052"
+เพชรบุรี,"482,193"
+ประจวบคีรีขันธ์,"550,678"
+เชียงใหม่,"1,784,370"
+ลำพูน,"402,011"
+ลำปาง,"728,964"
+อุตรดิตถ์,"448,745"
+แพร่,"437,350"
+น่าน,"476,727"
+พะเยา,"467,356"
+เชียงราย,"1,295,026"
+แม่ฮ่องสอน,"284,549"
+นครสวรรค์,"1,040,308"
+อุทัยธานี,"325,868"
+กำแพงเพชร,"714,118"
+ตาก,"670,265"
+สุโขทัย,"587,883"
+พิษณุโลก,"849,481"
+พิจิตร,"532,310"
+เพชรบูรณ์,"981,940"
+นครราชสีมา,"2,633,207"
+บุรีรัมย์,"1,581,184"
+สุรินทร์,"1,378,221"
+ศรีสะเกษ,"1,458,580"
+อุบลราชธานี,"1,866,697"
+ยโสธร,"534,500"
+ชัยภูมิ,"1,124,924"
+อำนาจเจริญ,"376,195"
+บึงกาฬ,"422,042"
+หนองบัวลำภู,"509,470"
+ขอนแก่น,"1,794,531"
+อุดรธานี,"1,567,983"
+เลย,"638,736"
+หนองคาย,"517,435"
+มหาสารคาม,"953,660"
+ร้อยเอ็ด,"1,298,640"
+กาฬสินธุ์,"977,175"
+สกลนคร,"1,146,936"
+นครพนม,"717,201"
+มุกดาหาร,"350,911"
+นครศรีธรรมราช,"1,550,721"
+กระบี่,"477,770"
+พังงา,"268,229"
+ภูเก็ต,"414,471"
+สุราษฎร์ธานี,"1,067,726"
+ระนอง,"194,372"
+ชุมพร,"509,208"
+สงขลา,"1,428,609"
+สตูล,"324,098"
+ตรัง,"640,574"
+พัทลุง,"523,077"
+ปัตตานี,"726,015"
+ยะลา,"538,602"
+นราธิวาส,"804,429"
