+++ bot_jobs/get_data.js
@@ -1,7 +1,7 @@
 const request = require('request')
 const csv = require('csv-parser')
 const fs = require('fs')
-/*
+
 request('https://data.go.th/dataset/8a956917-436d-4afd-a2d4-59e4dd8e906e/resource/be19a8ad-ab48-4081-b04a-8035b5b2b8d6/download/dataset.csv', (err, response, body) => {
     if (!err && response.statusCode == 200) {
         var dataset = body
@@ -18,14 +18,14 @@ request('https://data.go.th/dataset/8a956917-436d-4afd-a2d4-59e4dd8e906e/resourc
         console.log(err)
     }
 })
-*/
+
 request('https://covid19.th-stat.com/json/covid19v2/getTimeline.json', (err, response, body) => {
     if (!err && response.statusCode == 200) {
         body = JSON.parse(body)
         body['Data'] = body['Data'].filter(s=>{
             return new Date(s['Date']) >= new Date('7/1/2020')
         })
-        fs.writeFileSync('../components/gis/data/national-timeseries.json', JSON.stringify(body));
+        fs.writeFileSync('../components/gis/data/national-timeseries.json', JSON.stringify(body,null,2));
         console.log('national stats downloaded')
     }
     else {
