--- bot_jobs/tests/.gitignore
@@ -1,3 +1,0 @@
-pdata-current.json
-provinces-current.json
-vaccines-current.json
