--- bot_jobs/tests/vaccines-template.json
@@ -1,1 +1,0 @@
-{"\u0e01\u0e23\u0e38\u0e07\u0e40\u0e17\u0e1e\u0e21\u0e2b\u0e32\u0e19\u0e04\u0e23": 11.9, "\u0e2a\u0e21\u0e38\u0e17\u0e23\u0e1b\u0e23\u0e32\u0e01\u0e32\u0e23": 5.42, "\u0e19\u0e19\u0e17\u0e1a\u0e38\u0e23\u0e35": 8.27, "\u0e1b\u0e17\u0e38\u0e21\u0e18\u0e32\u0e19\u0e35": 3.78, "\u0e1e\u0e23\u0e30\u0e19\u0e04\u0e23\u0e28\u0e23\u0e35\u0e2d\u0e22\u0e38\u0e18\u0e22\u0e32": 2.1, "\u0e2d\u0e48\u0e32\u0e07\u0e17\u0e2d\u0e07": 3.43, "\u0e25\u0e1e\u0e1a\u0e38\u0e23\u0e35": 2.16, "\u0e2a\u0e34\u0e07\u0e2b\u0e4c\u0e1a\u0e38\u0e23\u0e35": 4.03, "\u0e0a\u0e31\u0e22\u0e19\u0e32\u0e17": 2.14, "\u0e2a\u0e23\u0e30\u0e1a\u0e38\u0e23\u0e35": 3.18, "\u0e0a\u0e25\u0e1a\u0e38\u0e23\u0e35": 5.42, "\u0e23\u0e30\u0e22\u0e2d\u0e07": 3.26, "\u0e08\u0e31\u0e19\u0e17\u0e1a\u0e38\u0e23\u0e35": 3.55, "\u0e15\u0e23\u0e32\u0e14": 3.7, "\u0e09\u0e30\u0e40\u0e0a\u0e34\u0e07\u0e40\u0e17\u0e23\u0e32": 4.13, "\u0e1b\u0e23\u0e32\u0e08\u0e35\u0e19\u0e1a\u0e38\u0e23\u0e35": 3.03, "\u0e19\u0e04\u0e23\u0e19\u0e32\u0e22\u0e01": 5.15, "\u0e2a\u0e23\u0e30\u0e41\u0e01\u0e49\u0e27": 3.69, "\u0e23\u0e32\u0e0a\u0e1a\u0e38\u0e23\u0e35": 1.98, "\u0e01\u0e32\u0e0d\u0e08\u0e19\u0e1a\u0e38\u0e23\u0e35": 1.58, "\u0e2a\u0e38\u0e1e\u0e23\u0e23\u0e13\u0e1a\u0e38\u0e23\u0e35": 2.32, "\u0e19\u0e04\u0e23\u0e1b\u0e10\u0e21": 2.29, "\u0e2a\u0e21\u0e38\u0e17\u0e23\u0e2a\u0e32\u0e04\u0e23": 12.88, "\u0e2a\u0e21\u0e38\u0e17\u0e23\u0e2a\u0e07\u0e04\u0e23\u0e32\u0e21": 5.12, "\u0e40\u0e1e\u0e0a\u0e23\u0e1a\u0e38\u0e23\u0e35": 4.9, "\u0e1b\u0e23\u0e30\u0e08\u0e27\u0e1a\u0e04\u0e35\u0e23\u0e35\u0e02\u0e31\u0e19\u0e18\u0e4c": 4.83, "\u0e40\u0e0a\u0e35\u0e22\u0e07\u0e43\u0e2b\u0e21\u0e48": 3.73, "\u0e25\u0e33\u0e1e\u0e39\u0e19": 2.13, "\u0e25\u0e33\u0e1b\u0e32\u0e07": 2.16, "\u0e2d\u0e38\u0e15\u0e23\u0e14\u0e34\u0e15\u0e16\u0e4c": 2.42, "\u0e41\u0e1e\u0e23\u0e48": 2.03, "\u0e19\u0e48\u0e32\u0e19": 2.02, "\u0e1e\u0e30\u0e40\u0e22\u0e32": 2.76, "\u0e40\u0e0a\u0e35\u0e22\u0e07\u0e23\u0e32\u0e22": 2.25, "\u0e41\u0e21\u0e48\u0e2e\u0e48\u0e2d\u0e07\u0e2a\u0e2d\u0e19": 2.65, "\u0e19\u0e04\u0e23\u0e2a\u0e27\u0e23\u0e23\u0e04\u0e4c": 1.8, "\u0e2d\u0e38\u0e17\u0e31\u0e22\u0e18\u0e32\u0e19\u0e35": 2.82, "\u0e01\u0e33\u0e41\u0e1e\u0e07\u0e40\u0e1e\u0e0a\u0e23": 2.1, "\u0e15\u0e32\u0e01": 8.91, "\u0e2a\u0e38\u0e42\u0e02\u0e17\u0e31\u0e22": 2.3, "\u0e1e\u0e34\u0e29\u0e13\u0e38\u0e42\u0e25\u0e01": 2.58, "\u0e1e\u0e34\u0e08\u0e34\u0e15\u0e23": 2.28, "\u0e40\u0e1e\u0e0a\u0e23\u0e1a\u0e39\u0e23\u0e13\u0e4c": 2.4, "\u0e19\u0e04\u0e23\u0e23\u0e32\u0e0a\u0e2a\u0e35\u0e21\u0e32": 3.72, "\u0e1a\u0e38\u0e23\u0e35\u0e23\u0e31\u0e21\u0e22\u0e4c": 6.38, "\u0e2a\u0e38\u0e23\u0e34\u0e19\u0e17\u0e23\u0e4c": 1.81, "\u0e28\u0e23\u0e35\u0e2a\u0e30\u0e40\u0e01\u0e29": 1.64, "\u0e2d\u0e38\u0e1a\u0e25\u0e23\u0e32\u0e0a\u0e18\u0e32\u0e19\u0e35": 1.78, "\u0e22\u0e42\u0e2a\u0e18\u0e23": 2.17, "\u0e0a\u0e31\u0e22\u0e20\u0e39\u0e21\u0e34": 2.12, "\u0e2d\u0e33\u0e19\u0e32\u0e08\u0e40\u0e08\u0e23\u0e34\u0e0d": 2.26, "\u0e1a\u0e36\u0e07\u0e01\u0e32\u0e2c": 2.57, "\u0e2b\u0e19\u0e2d\u0e07\u0e1a\u0e31\u0e27\u0e25\u0e33\u0e20\u0e39": 1.95, "\u0e02\u0e2d\u0e19\u0e41\u0e01\u0e48\u0e19": 1.98, "\u0e2d\u0e38\u0e14\u0e23\u0e18\u0e32\u0e19\u0e35": 2.35, "\u0e40\u0e25\u0e22": 2.4, "\u0e2b\u0e19\u0e2d\u0e07\u0e04\u0e32\u0e22": 3.85, "\u0e21\u0e2b\u0e32\u0e2a\u0e32\u0e23\u0e04\u0e32\u0e21": 1.76, "\u0e23\u0e49\u0e2d\u0e22\u0e40\u0e2d\u0e47\u0e14": 1.74, "\u0e01\u0e32\u0e2c\u0e2a\u0e34\u0e19\u0e18\u0e38\u0e4c": 1.75, "\u0e2a\u0e01\u0e25\u0e19\u0e04\u0e23": 2.22, "\u0e19\u0e04\u0e23\u0e1e\u0e19\u0e21": 2.52, "\u0e21\u0e38\u0e01\u0e14\u0e32\u0e2b\u0e32\u0e23": 4.69, "\u0e19\u0e04\u0e23\u0e28\u0e23\u0e35\u0e18\u0e23\u0e23\u0e21\u0e23\u0e32\u0e0a": 2.34, "\u0e01\u0e23\u0e30\u0e1a\u0e35\u0e48": 5.35, "\u0e1e\u0e31\u0e07\u0e07\u0e32": 8.62, "\u0e20\u0e39\u0e40\u0e01\u0e47\u0e15": 44.0, "\u0e2a\u0e38\u0e23\u0e32\u0e29\u0e0e\u0e23\u0e4c\u0e18\u0e32\u0e19\u0e35": 6.62, "\u0e23\u0e30\u0e19\u0e2d\u0e07": 12.68, "\u0e0a\u0e38\u0e21\u0e1e\u0e23": 2.37, "\u0e2a\u0e07\u0e02\u0e25\u0e32": 3.12, "\u0e2a\u0e15\u0e39\u0e25": 3.61, "\u0e15\u0e23\u0e31\u0e07": 2.43, "\u0e1e\u0e31\u0e17\u0e25\u0e38\u0e07": 2.31, "\u0e1b\u0e31\u0e15\u0e15\u0e32\u0e19\u0e35": 2.35, "\u0e22\u0e30\u0e25\u0e32": 2.41, "\u0e19\u0e23\u0e32\u0e18\u0e34\u0e27\u0e32\u0e2a": 2.99}
