--- bot_jobs/th-census-age-group.csv
@@ -1,83 +1,0 @@
-"province","<20","20-60",">60","total","PROV_CODE","PROV_NAME","clean-name","fid"
-"ทั่วราชอาณาจักร",14709329,38664947,11633188,66186727,,,,
-"กรุงเทพมหานคร",1058164,3269617,1108912,5588222,"10","BANGKOK","BANGKOK",1
-"ภาคกลาง",3767547,10136304,3005542,17255105,,,,
-"สมุทรปราการ",293327,818770,214241,1351479,"11","SAMUT PRAKAN","SAMUTPRAKAN",2
-"นนทบุรี",249192,765766,239510,1276745,"12","NONTHABURI","NONTHABURI",3
-"ปทุมธานี",261017,724780,175360,1176412,"13","PATHUM THANI","PATHUMTHANI",4
-"พระนครศรีอยุธยา",175590,484196,155143,819088,"14","PHRA NAKHON SI AYUTTHAYA","PHRANAKHONSIAYUTTHAYA",5
-"อ่างทอง",55286,158028,62315,276584,"15","ANG THONG","ANGTHONG",6
-"ลพบุรี",152305,437722,146807,742928,"16","LOP BURI","LOPBURI",7
-"สิงห์บุรี",39618,115775,49921,205898,"17","SING BURI","SINGBURI",8
-"ชัยนาท",63390,183440,74901,322477,"18","CHAI NAT","CHAINAT",9
-"สระบุรี",143452,378901,112257,643828,"19","SARABURI","SARABURI",10
-"ชลบุรี",377493,938836,216726,1566885,"20","CHON BURI","CHONBURI",11
-"ระยอง",183262,447725,99566,741524,"21","RAYONG","RAYONG",12
-"จันทบุรี",118345,312277,99809,535559,"22","CHANTHABURI","CHANTHABURI",13
-"ตราด",49227,127773,41300,228536,"23","TRAT","TRAT",14
-"ฉะเชิงเทรา",164765,423293,126675,720718,"24","CHACHOENGSAO","CHACHOENGSAO",15
-"ปราจีนบุรี",113724,293265,84339,493670,"25","PRACHIN BURI","PRACHINBURI",16
-"นครนายก",56325,149055,52395,260081,"26","NAKHON NAYOK","NAKHONNAYOK",17
-"สระแก้ว",137028,332187,86405,560925,"27","SA KAEO","SAKAEO",18
-"ราชบุรี",180988,492614,167016,869313,"70","RATCHABURI","RATCHABURI",56
-"กาญจนบุรี",193421,480638,137572,891976,"71","KANCHANABURI","KANCHANABURI",57
-"สุพรรณบุรี",175424,483519,175302,838628,"72","SUPHAN BURI","SUPHANBURI",58
-"นครปฐม",195815,547085,164646,920729,"73","NAKHON PATHOM","NAKHONPATHOM",59
-"สมุทรสาคร",125202,332269,91464,586199,"74","SAMUT SAKHON","SAMUTSAKHON",60
-"สมุทรสงคราม",36022,109368,44569,192052,"75","SAMUT SONGKHRAM","SAMUTSONGKHRAM",61
-"เพชรบุรี",102363,280466,93990,482193,"76","PHETCHABURI","PHETCHABURI",62
-"ประจวบคีรีขันธ์",124966,318556,93313,550678,"77","PRACHUAP KHIRI KHAN","PRACHUAPKHIRIKHAN",63
-"ภาคเหนือ",2415822,6720480,2390951,12027271,,,,
-"เชียงใหม่",339501,933947,349970,1784370,"50","CHIANG MAI","CHIANGMAI",39
-"ลำพูน",70510,231147,96805,402011,"51","LAMPHUN","LAMPHUN",40
-"ลำปาง",119321,427957,177899,728964,"52","LAMPANG","LAMPANG",41
-"อุตรดิตถ์",86477,261172,99336,448745,"53","UTTARADIT","UTTARADIT",42
-"แพร่",73886,257165,104954,437350,"54","PHRAE","PHRAE",43
-"น่าน",94614,281559,98389,476727,"55","NAN","NAN",44
-"พะเยา",82846,276970,104517,467356,"56","PHAYAO","PHAYAO",45
-"เชียงราย",251578,676593,238087,1295026,"57","CHIANG RAI","CHIANGRAI",46
-"แม่ฮ่องสอน",70355,135344,34458,284549,"58","MAE HONG SON","MAEHONGSON",47
-"นครสวรรค์",214800,602818,218041,1040308,"60","NAKHON SAWAN","NAKHONSAWAN",48
-"อุทัยธานี",71292,186252,67296,325868,"61","UTHAI THANI","UTHAITHANI",49
-"กำแพงเพชร",160128,419316,131769,714118,"62","KAMPHAENG PHET","KAMPHAENGPHET",50
-"ตาก",163196,298637,79826,670265,"63","TAK","TAK",51
-"สุโขทัย",117225,345008,124307,587883,"64","SUKHOTHAI","SUKHOTHAI",52
-"พิษณุโลก",176607,501053,167018,849481,"65","PHITSANULOK","PHITSANULOK",53
-"พิจิตร",109304,306589,113617,532310,"66","PHICHIT","PHICHIT",54
-"เพชรบูรณ์",214182,578953,184662,981940,"67","PHETCHABUN","PHETCHABUN",55
-"ภาคตะวันออกเฉียงเหนือ",4924326,13152166,3685641,21848228,,,,
-"นครราชสีมา",582374,1566988,473644,2633207,"30","NAKHON RATCHASIMA","NAKHONRATCHASIMA",19
-"บุรีรัมย์",378608,936557,261669,1581184,"31","BURI RAM","BURIRAM",20
-"สุรินทร์",324916,816020,233749,1378221,"32","SURIN","SURIN",21
-"ศรีสะเกษ",334400,877873,242643,1458580,"33","SI SA KET","SISAKET",22
-"อุบลราชธานี",445677,1120158,289179,1866697,"34","UBON RATCHATHANI","UBONRATCHATHANI",23
-"ยโสธร",112327,326222,94795,534500,"35","YASOTHON","YASOTHON",24
-"ชัยภูมิ",245540,664698,211898,1124924,"36","CHAIYAPHUM","CHAIYAPHUM",25
-"อำนาจเจริญ",84041,228489,62492,376195,"37","AMNAT CHAROEN","AMNATCHAROEN",26
-"บึงกาฬ",106357,253946,60389,422042,"38","BUENG KAN","BUENGKAN",27
-"หนองบัวลำภู",117243,310594,80433,509470,"39","NONG BUA LAMPHU","NONGBUALAMPHU",28
-"ขอนแก่น",379036,1082245,326053,1794531,"40","KHON KAEN","KHONKAEN",29
-"อุดรธานี",354638,957456,245343,1567983,"41","UDON THANI","UDONTHANI",30
-"เลย",143188,374055,116590,638736,"42","LOEI","LOEI",31
-"หนองคาย",115652,311049,84533,517435,"43","NONG KHAI","NONGKHAI",32
-"มหาสารคาม",197655,583085,170986,953660,"44","MAHA SARAKHAM","MAHASARAKHAM",33
-"ร้อยเอ็ด",268603,794319,232664,1298640,"45","ROI ET","ROIET",34
-"กาฬสินธุ์",210260,600479,164502,977175,"46","KALASIN","KALASIN",35
-"สกลนคร",271641,700337,171997,1146936,"47","SAKON NAKHON","SAKONNAKHON",36
-"นครพนม",170820,434275,109302,717201,"48","NAKHON PHANOM","NAKHONPHANOM",37
-"มุกดาหาร",81350,213321,52780,350911,"49","MUKDAHAN","MUKDAHAN",38
-"ภาคใต้",2543470,5386380,1442142,9467901,,,,
-"นครศรีธรรมราช",368477,899055,276533,1550721,"80","NAKHON SI THAMMARAT","NAKHONSITHAMMARAT",64
-"กระบี่",137806,276554,59700,477770,"81","KRABI","KRABI",65
-"พังงา",65129,151891,46146,268229,"82","PHANGNGA","PHANGNGA",66
-"ภูเก็ต",109365,243034,48262,414471,"83","PHUKET","PHUKET",67
-"สุราษฎร์ธานี",268445,622870,164938,1067726,"84","SURAT THANI","SURATTHANI",68
-"ระนอง",45206,105132,28289,194372,"85","RANONG","RANONG",69
-"ชุมพร",115469,296787,91744,509208,"86","CHUMPHON","CHUMPHON",70
-"สงขลา",368274,813755,229680,1428609,"90","SONGKHLA","SONGKHLA",71
-"สตูล",95931,183430,42848,324098,"91","SATUN","SATUN",72
-"ตรัง",157583,374589,105627,640574,"92","TRANG","TRANG",73
-"พัทลุง",119473,302266,100108,523077,"93","PHATTHALUNG","PHATTHALUNG",74
-"ปัตตานี",247677,386170,88552,726015,"94","PATTANI","PATTANI",75
-"ยะลา",180257,288780,65546,538602,"95","YALA","YALA",76
-"นราธิวาส",264378,442067,94169,804429,"96","NARATHIWAT","NARATHIWAT",77
